requirements.txt
flask==2.3.3
sentence-transformers==2.2.2
transformers==4.36.2
torch==2.1.2+cu121
torchvision==0.16.2+cu121
torchaudio==2.1.2+cu121
--extra-index-url https://download.pytorch.org/whl/cu121
PyMuPDF==1.23.9
python-docx==1.1.0
requests==2.31.0
numpy==1.24.4
scikit-learn==1.3.2
python-dotenv==1.0.0
accelerate==0.24.1
huggingface-hub==0.19.4 